requirements.txt
asgiref==3.5.2
certifi==2020.12.5
chardet==4.0.0
Django==4.1.3
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.2
urllib3==1.26.5
workos>=5.37.0
python-dotenv
django_feather==0.1.0
